+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning].
 
 - RandomSkunk.Results:
     - Add optional `onNoneSelector` parameter to `Maybe<T>.Select` and `Maybe<T>.SelectMany` methods.
+    - Add implicit conversion operator methods from `T` to `Result<T>` and `Maybe<T>`.
 
 ## [1.0.0-alpha23] - 2023-01-03